Allocate space for arrays when deserializing JSON
Previously, when using dynamic mode and deserializing an incoming JSON
message from MQTT, the library would allocate space based on the number
of colons - i.e. the number of object key-value pairs. However, if there
are arrays in the JSON message (such as when a shared attribute gets
deleted), that wouldn't allocate enough space.